requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
numpy==1.21.5
packaging==21.3
panda==0.3.1
pandas==1.3.5
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
